// app/models/events.ts
export enum EventAction {
  Create = 'Create',
  Update = 'Update',
  Delete = 'Delete',
  Login = 'Login',
  logout = 'logout',
}

export enum EventDomain {
  Tenant = 'Tenant',
  PropertyType = 'PropertyType',
  Insurance = 'Insurance',
  RoofType = 'RoofType',
  InsuranceItem = 'InsuranceItem',
  Outgoing = 'Outgoing',
  Parking = 'Parking',
  ParkingType = 'ParkingType',
  Plot = 'Plot',
  ComparablePlot = 'ComparablePlot',
  PlotAndComparable = 'PlotAndComparable',
  Image = 'Image',
  Client = 'Client',
  Valuer = 'Valuer',
  Instruction = 'Instruction',
  Grc = 'Grc',
  GrcFee = 'GrcFee',
  GrcDepr = 'GrcDepr',
  Mv = 'Mv',
  Password = 'Password',
  User = 'User',
  Company = 'Company',
  StoredValue = 'StoredValue',
  SavedConfig = 'SavedConfig',
  Suburb = 'Suburb',
  UserGroup = 'UserGroup',
  RptTemplate = 'RptTemplate',
  ReportTemplate = "ReportTemplate",
  Notification = "Notification",
  ReportHeader = "ReportHeader",
  SubHeader = "SubHeader",
  BodyContent = "BodyContent",
  CompanyImage = "CompanyImage",
  ValuationsHistory = "Valuations History",
  Attachment = "Attachment",
}